requirements.txt
alembic==1.3.2
bleach==3.1.0
blinker==1.4
Click==7.0
Flask==1.1.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
Flask-Uploads==0.2.1
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10.3
Mako==1.1.1
Markdown==3.1.1
MarkupSafe==1.1.1
python-dateutil==2.8.1
python-editor==1.0.4
python-slugify==4.0.0
six==1.14.0
SQLAlchemy==1.3.12
text-unidecode==1.3
webencodings==0.5.1
Werkzeug==0.16.0
WTForms==2.2.1
